refactor(e2e): split shared/panels god helpers into topic libraries
Extract residual implementations from e2e_test_shared and e2e_test_shared_panels into focused topic modules while keeping umbrella re-exports for AC1 compatibility (Refs #4075 Slice A).
