+++ packages/colonizethis_app_e2e_support/lib/e2e_test_shared.dart
@@ -1,16 +1,5 @@
-import 'package:colonizethis_app_fixtures/config/ct_e2e.dart';
-import 'package:colonizethis_app/features/game/widgets/dialogue/game_start_intro_overlay.dart';
-import 'package:colonizethis_app/widgets/ct_nine_patch_button.dart';
-import 'package:colonizethis_app/features/game/widgets/units/civilian/civilian_units_panel.dart'
-    show CivilianUnitRowCard;
-import 'package:colonizethis_app/widgets/ct_dialog_shell.dart';
-import 'package:flutter/material.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter_test/flutter_test.dart';
-
-import 'package:colonizethis_app_e2e_support/e2e_test_shared_dismiss_alert_dialog.dart';
-import 'package:colonizethis_app_e2e_support/e2e_test_shared_dismiss_ct_dialog_shell_broad_sweep.dart';
-import 'package:colonizethis_app_e2e_support/e2e_test_shared_dismiss_generic_ok.dart';
-import 'package:colonizethis_app_e2e_support/e2e_test_shared_dismiss_snackbar.dart';
 
 export 'e2e_test_shared_bundled_explore_failure.dart';
 export 'e2e_test_shared_bundled_explore_retry.dart';
@@ -22,6 +11,7 @@ export 'e2e_test_shared_dismiss_ct_dialog_shell_broad_sweep.dart';
 export 'e2e_test_shared_dismiss_ct_dialog_shell_escalation.dart';
 export 'e2e_test_shared_dismiss_generic_ok.dart';
 export 'e2e_test_shared_dismiss_snackbar.dart';
+export 'e2e_test_shared_dismiss_transient_ui.dart';
 export 'e2e_test_shared_expansion_tile.dart';
 export 'e2e_test_shared_final_naval_reach_check.dart';
 export 'e2e_test_shared_first_fleet_move.dart';
@@ -38,6 +28,11 @@ export 'e2e_test_shared_panel_text_match.dart';
 export 'e2e_test_shared_panels.dart';
 export 'e2e_test_shared_region_tabs.dart';
 export 'e2e_test_shared_standard_scenario_opener.dart';
+export 'e2e_test_shared_intro.dart';
+export 'e2e_test_shared_bottom_sheet.dart';
+export 'e2e_test_shared_hit_testable.dart';
+export 'e2e_test_shared_civilian_taps.dart';
+export 'e2e_test_shared_move_dialog_finders.dart';
 
 /// Next interval after an idle poll pump in E2E busy-wait loops (25→50→75→100 ms).
 /// Aligns with [e2eWaitUntilFound] backoff (`SPEC/program/e2e-integration-tests.md`, #2336).
@@ -68,176 +63,6 @@ class E2ePerfLog {
       'E2E_TIMING|test=$testName|phase=$phase|ms=${elapsed.inMilliseconds}$metaPart',
     );
   }
-}
-
-/// True while the game-start intro still shows its blocking shell or spinner.
-///
-/// [GameStartIntroOverlay] stays mounted after dismissal; only the blocking
-/// [CtDialogShell] / [GameStartIntroLoadingIndicator] indicate UI capture.
-bool e2eGameStartIntroBlocksUi(WidgetTester tester) {
-  if (find.byType(GameStartIntroLoadingIndicator).evaluate().isNotEmpty) {
-    return true;
-  }
-  if (find.byType(GameStartIntroOverlay).evaluate().isEmpty) {
-    return false;
-  }
-  return find
-      .descendant(
-        of: find.byType(GameStartIntroOverlay),
-        matching: find.byType(CtDialogShell),
-      )
-      .evaluate()
-      .isNotEmpty;
-}
-
-/// Default phase label emitted by [e2eAdvanceGameStartIntroUntilDismissed]
-/// when a caller does not override [E2ePerfLog] attribution.
-///
-/// The constant is exposed so widget-test pins and downstream perf-marker
-/// scrapers (for example the GitHub #2336 AC8 baseline timing pipeline via
-/// `tool/run_e2e_timing.sh` / `tool/compare_e2e_timing.sh`) can refer to the
-/// canonical label by name instead of hard-coding the literal string. Mirrors
-/// the [kE2eDefaultWaitForMapHudPhase] convention introduced for the map-HUD
-/// bootstrap wait in PR #2960.
-const String kE2eDefaultAdvanceGameStartIntroPhase =
-    'advance_game_start_intro_until_dismissed';
-
-/// Counter name bumped on every [e2eAdvanceGameStartIntroUntilDismissed]
-/// iteration so a hung intro dismissal surfaces as an attributable counter
-/// spike instead of a silent 15 s wall-clock burn (Refs GitHub #2336 AC8 /
-/// AC10). Mirrors [kE2eWaitForMapHudIterationsCounter] which carries the same
-/// "iteration tally surfaces hangs as a counter spike" contract for the
-/// downstream map-HUD wait.
-const String kE2eAdvanceGameStartIntroIterationsCounter =
-    'advance_game_start_intro_until_dismissed_iterations';
-
-/// Yarn intro control labels tried in order each loop iteration (GitHub #2336).
-///
-/// The production `game_start_intro` node (one narrative line then
-/// `-> I shall.`) is collapsed by `CtDialogueView` into a **single** step whose
-/// only control is the Yarn option label `I shall.` (Refs #3628): there is no
-/// intermediate Continue line step, so one tap of `I shall.` dismisses the
-/// intro. The legacy generic `Continue` label is retained as a defensive
-/// fallback (e.g. the asset-load error shell) and tried only after `I shall.`.
-const List<String> kE2eGameStartIntroControlLabels = ['I shall.', 'Continue'];
-
-/// Per-control post-tap settle budget after an intro button tap.
-///
-/// Intermediate controls (for example **Continue** before **I shall.**) do not
-/// dismiss the blocking shell; the legacy 5 s [e2ePumpUntilConditionOrIdle]
-/// cap per tap therefore burned ~5 s on every bootstrap before the next label
-/// could be tried. A shorter cap keeps frame settlement without blocking the
-/// multi-label pass (Refs GitHub #2336 AC5 / bootstrap wall-clock).
-const Duration kE2eDefaultIntroControlPostTapSettleTimeout = Duration(
-  milliseconds: 500,
-);
-
-/// Advances yarn intro lines/choices until the overlay no longer blocks taps.
-///
-/// The spinner / no-tap-target branches previously each paid a fixed 50 ms
-/// pump per loop iteration. Both now share an [e2eAdaptivePollRampAfterIdle]
-/// idle pump (25 → 50 → 75 → 100 ms cap) so a long spinner stretch settles
-/// with adaptive backoff instead of constant 50 ms frames. The poll cadence
-/// is reset to 25 ms whenever a tap advances the overlay or the loading
-/// indicator clears, mirroring the prepump-free panel openers landed in this
-/// PR. Refs GitHub #2336 AC5 / pump-reduction.
-///
-/// **Multi-label pass:** each loop iteration tries every label in
-/// [kE2eGameStartIntroControlLabels] without breaking after the first tap.
-/// Intermediate controls (for example **Continue** before **I shall.**) keep
-/// the blocking shell mounted; the per-control post-tap settle uses
-/// [kE2eDefaultIntroControlPostTapSettleTimeout] (500 ms) instead of the
-/// legacy 5 s cap so bootstrap does not burn ~5 s waiting for full dismissal
-/// before the next control is tapped.
-///
-/// Perf attribution (Refs GitHub #2336 AC8 / baseline measurement):
-///
-/// - When [perf] is non-null, emits one `E2E_TIMING|phase=[phaseName]` line
-///   on every return path with a `result=...` meta tag distinguishing
-///   `result=already_dismissed` (entry-iteration short-circuit; counter
-///   value `1`), `result=advanced` (intro stopped blocking on a later
-///   iteration after one or more pumps or taps), and `result=timeout`
-///   (overall-cap fail path). The phase label defaults to
-///   [kE2eDefaultAdvanceGameStartIntroPhase].
-/// - Bumps the [kE2eAdvanceGameStartIntroIterationsCounter] counter once per
-///   loop iteration (including the iteration that returns success) so a
-///   hung intro dismissal surfaces as a counter spike instead of a silent
-///   wall-clock burn. The counter is incremented before the early-exit check
-///   for that iteration so the `result=already_dismissed` short-circuit
-///   reports `value=1`.
-/// - With `perf: null` (the default for the existing widget-test pins and any
-///   opt-out callers) the helper emits no `E2E_TIMING` / `E2E_COUNTER` lines.
-Future<void> e2eAdvanceGameStartIntroUntilDismissed(
-  WidgetTester tester, {
-  E2ePerfLog? perf,
-  Duration timeout = const Duration(seconds: 15),
-  String phaseName = kE2eDefaultAdvanceGameStartIntroPhase,
-}) async {
-  final sw = Stopwatch()..start();
-  final deadline = DateTime.now().add(timeout);
-  var idlePollMs = 25;
-  var iterations = 0;
-  while (DateTime.now().isBefore(deadline)) {
-    iterations += 1;
-    perf?.bumpCounter(
-      kE2eAdvanceGameStartIntroIterationsCounter,
-      meta: 'phase=$phaseName',
-    );
-    if (!e2eGameStartIntroBlocksUi(tester)) {
-      perf?.timing(
-        phaseName,
-        sw.elapsed,
-        meta: iterations == 1 ? 'result=already_dismissed' : 'result=advanced',
-      );
-      return;
-    }
-    if (find.byType(GameStartIntroLoadingIndicator).evaluate().isNotEmpty) {
-      await tester.pump(Duration(milliseconds: idlePollMs));
-      idlePollMs = e2eAdaptivePollRampAfterIdle(idlePollMs);
-      continue;
-    }
-    final overlay = find.byType(GameStartIntroOverlay);
-    var tapped = false;
-    for (final label in kE2eGameStartIntroControlLabels) {
-      if (!e2eGameStartIntroBlocksUi(tester)) {
-        break;
-      }
-      final control = find
-          .descendant(of: overlay, matching: find.text(label))
-          .hitTestable();
-      if (control.evaluate().isEmpty) {
-        continue;
-      }
-      await tester.tap(control.first, warnIfMissed: false);
-      await e2ePumpUntilConditionOrIdle(
-        tester,
-        () => !e2eGameStartIntroBlocksUi(tester),
-        timeout: kE2eDefaultIntroControlPostTapSettleTimeout,
-        perf: perf,
-        phaseName: 'pump_until_intro_advance_after_$label',
-      );
-      tapped = true;
-      idlePollMs = 25;
-    }
-    if (!tapped) {
-      await tester.pump(Duration(milliseconds: idlePollMs));
-      idlePollMs = e2eAdaptivePollRampAfterIdle(idlePollMs);
-    }
-  }
-  if (e2eGameStartIntroBlocksUi(tester)) {
-    perf?.timing(phaseName, sw.elapsed, meta: 'result=timeout');
-    fail(
-      'Timed out after ${timeout.inSeconds}s advancing game start intro. '
-      'Last exception: ${tester.takeException()}',
-    );
-  }
-  // Implicit success: the wall-clock deadline elapsed but the intro stopped
-  // blocking on the very last iteration (rare, but observable in CI where
-  // the loop's deadline check races a final tap). Attribute as
-  // `result=advanced` — the helper still returned without the fail-path —
-  // so the AC8 timing pipeline does not silently bucket the success into
-  // `result=timeout`.
-  perf?.timing(phaseName, sw.elapsed, meta: 'result=advanced');
 }
 
 Future<void> e2ePumpFor(WidgetTester tester, Duration total) async {
@@ -279,808 +104,6 @@ Future<void> e2ePumpUntilFinderEmpty(
     }
     stepMs = e2eAdaptivePollRampAfterIdle(stepMs);
   }
-}
-
-/// Default cap for bottom-sheet close polling (GitHub #2336).
-const Duration kE2eDefaultBottomSheetCloseTimeout = Duration(seconds: 5);
-
-/// Closes an open [BottomSheet] by issuing a single back-route pop and polling
-/// until the sheet leaves the widget tree.
-///
-/// Shared by full-turn and fleet E2E; [overallTimeout] defaults to
-/// [kE2eDefaultBottomSheetCloseTimeout] (previous per-file caps).
-///
-/// **Why pop once:** the previous implementation called
-/// [tester.binding.handlePopRoute] on every poll iteration, even while the
-/// dismiss animation was already running. A single pop initiates the dismiss;
-/// subsequent calls are wasted work and prevent the loop from short-circuiting
-/// on the post-dismiss frame. The replacement pops once, then delegates to
-/// [e2ePumpUntilConditionOrIdle] so the loop exits the moment the sheet is
-/// gone. If the first pop is dropped (rare; route stack stale), a single
-/// retry pop is issued before falling back to the failure path. Refs
-/// GitHub #2336 (`pump-reduction` slice).
-Future<void> e2eCloseBottomSheet(
-  WidgetTester tester, {
-  E2ePerfLog? perf,
-  Duration overallTimeout = kE2eDefaultBottomSheetCloseTimeout,
-}) async {
-  perf?.bumpCounter('close_bottom_sheet_calls');
-  bool anyPanelOpen() => find.byType(BottomSheet).evaluate().isNotEmpty;
-
-  if (!anyPanelOpen()) {
-    return;
-  }
-
-  final sw = Stopwatch()..start();
-  await tester.binding.handlePopRoute();
-  final firstWindow = overallTimeout < const Duration(seconds: 2)
-      ? overallTimeout
-      : const Duration(seconds: 2);
-  if (await e2ePumpUntilConditionOrIdle(
-    tester,
-    () => !anyPanelOpen(),
-    timeout: firstWindow,
-    perf: perf,
-    phaseName: 'pump_until_bottom_sheet_closed_after_pop',
-  )) {
-    perf?.timing('close_bottom_sheet', sw.elapsed);
-    return;
-  }
-  if (!anyPanelOpen()) {
-    perf?.timing('close_bottom_sheet', sw.elapsed);
-    return;
-  }
-  // First pop may be dropped if the route stack changed mid-frame (e.g. an
-  // overlay raced the pop). Retry once, then wait out the remaining budget.
-  await tester.binding.handlePopRoute();
-  final remaining = overallTimeout - sw.elapsed;
-  if (remaining > Duration.zero) {
-    await e2ePumpUntilConditionOrIdle(
-      tester,
-      () => !anyPanelOpen(),
-      timeout: remaining,
-      perf: perf,
-      phaseName: 'pump_until_bottom_sheet_closed_after_retry',
-    );
-  }
-  if (!anyPanelOpen()) {
-    perf?.timing('close_bottom_sheet', sw.elapsed);
-    return;
-  }
-
-  fail(
-    'Timed out after ${overallTimeout.inSeconds}s closing bottom sheet; '
-    'panels remained visible',
-  );
-}
-
-/// Ensures [trigger] is visible (best-effort) and taps its first
-/// hit-testable element, falling back to the raw [trigger] when no
-/// hit-testable element resolves.
-///
-/// Encodes the defensive rail/marker tap sequence shared by
-/// [e2eOpenCivilianPanel], [e2eOpenNavalPanel], and [e2eOpenProductionPanel]
-/// so each panel opener gains uniform off-screen-trigger resilience
-/// (Refs GitHub #2336 AC1 / AC2 / AC10; PR #2555 history — the naval rail
-/// tap could land off-target until `ensureVisible` + hit-testable resolve
-/// was added inline, while the civilian opener kept the brittle raw-tap
-/// path that drops the tap when the rail button is rendered but pushed
-/// outside the viewport by a transient overlay).
-///
-/// Contract:
-///
-/// - When [trigger] resolves to zero elements, returns `false` without
-///   tapping. The caller remains responsible for the upstream presence
-///   check; the panel-opener `tryOpen` closures already gate on
-///   `trigger.evaluate().isNotEmpty` before calling, so this no-op branch
-///   is a safety net rather than the primary path.
-/// - Calls [WidgetTester.ensureVisible] inside a `try`/`catch (_)` so an
-///   `ensureVisible` failure (for example when the trigger is not in any
-///   `Scrollable`) does not throw past the helper.
-/// - Resolves the tap target via `trigger.hitTestable()`; falls back to
-///   the raw [trigger] when no element is hit-testable so the tap still
-///   fires from the same canonical position the legacy opener bodies
-///   used.
-/// - Taps the first resolved element with `warnIfMissed: false` to match
-///   the existing panel-opener contract.
-/// - Returns `true` when the tap was issued.
-Future<bool> e2eEnsureVisibleAndTapHitTestable(
-  WidgetTester tester,
-  Finder trigger,
-) async {
-  if (trigger.evaluate().isEmpty) {
-    return false;
-  }
-  try {
-    await tester.ensureVisible(trigger);
-  } catch (_) {}
-  final hit = trigger.hitTestable();
-  final target = hit.evaluate().isNotEmpty ? hit : trigger;
-  await tester.tap(target.first, warnIfMissed: false);
-  return true;
-}
-
-/// Awaits one of [primary] / [secondary] becoming hit-testable before the
-/// outer panel-opener loop attempts a [e2eEnsureVisibleAndTapHitTestable]
-/// tap, defending against transient overlays that cover the rail/marker
-/// trigger.
-///
-/// Lifts the inline rail/marker hit-testable wait that the naval opener
-/// has carried since PR #2555 (`wait_until_naval_rail_hit_testable` /
-/// `wait_until_naval_marker_hit_testable` phases) into a single shared
-/// primitive so [e2eOpenCivilianPanel], [e2eOpenNavalPanel], and
-/// [e2eOpenProductionPanel] gain identical pre-tap settle semantics.
-/// Before this lift, the civilian opener tapped a rail that might not be
-/// hit-testable (relying on `e2eEnsureVisibleAndTapHitTestable`'s
-/// best-effort scroll-into-view) and let the outer adaptive-poll loop
-/// retry on miss; production took the same shape. The naval opener
-/// instead waited up to 5 s for one of `[empireRailButton, markerButton]`
-/// to become hit-testable before tapping. Refs GitHub #2336 AC1 / AC10
-/// (deferred slice from PR #2782).
-///
-/// Contract:
-///
-/// - Short-circuits with no pump and no perf event when either [primary]
-///   or [secondary] is already hit-testable. The fast-path keeps the
-///   no-overlay common case at byte-equivalent cost vs the pre-#2555
-///   civilian/production openers.
-/// - Otherwise delegates to [e2eWaitUntilAnyFinderHitTestable] with the
-///   provided [timeout] and [phaseName]. That helper polls with adaptive
-///   backoff (25 → 500 ms cap) and emits `result=found{,_immediate,_at_
-///   timeout}` / `result=timeout` perf timings on [perf]. On final
-///   timeout it fails via [fail] with the same diagnostic message the
-///   inline naval opener has surfaced since PR #2555 (`Timed out after
-///   ${timeout.inSeconds}s waiting for any of ...`), so a regression that
-///   left the trigger permanently obscured fails with a useful message
-///   inside the inner wait rather than silently consuming the outer
-///   opener loop's full budget.
-/// - When [secondary] is `null`, only [primary] is polled (production
-///   opener path — no `markerButton` concept).
-Future<void> e2eAwaitPanelOpenerRailHitTestable(
-  WidgetTester tester, {
-  required Finder primary,
-  Finder? secondary,
-  Duration timeout = const Duration(seconds: 5),
-  E2ePerfLog? perf,
-  required String phaseName,
-}) async {
-  if (primary.hitTestable().evaluate().isNotEmpty) {
-    return;
-  }
-  if (secondary != null && secondary.hitTestable().evaluate().isNotEmpty) {
-    return;
-  }
-  final finders = <Finder>[primary, ?secondary];
-  await e2eWaitUntilAnyFinderHitTestable(
-    tester,
-    finders,
-    timeout: timeout,
-    perf: perf,
-    phaseName: phaseName,
-  );
-}
-
-/// Civilian work-menu labels surfaced after tapping an `Assign` button in
-/// the civilian panel (`Build improvement`, `Prospect`, `Explore`).
-///
-/// Single source of truth consumed by [e2eAwaitCivilianWorkMenuMounted] (and
-/// transitively by [e2eTapFirstAssignInCivilianPanel] and
-/// [e2eTapAssignOnCivilianRowWithTitle]) so a label drift in production
-/// surfaces in one place. Pinned by the widget test in
-/// `app/test/e2e_await_civilian_work_menu_mounted_test.dart` to keep the
-/// label set deterministic across SDK upgrades. Refs GitHub #2336 AC1 / AC2 /
-/// Bottleneck 6.
-const List<String> kE2eCivilianWorkMenuLabels = <String>[
-  'Build improvement',
-  'Prospect',
-  'Explore',
-];
-
-/// Default timeout for [e2eAwaitCivilianWorkMenuMounted]. Matches the legacy
-/// pre-lift inline `Duration(seconds: 5)` budget shared by
-/// [e2eTapFirstAssignInCivilianPanel] and [e2eTapAssignOnCivilianRowWithTitle]
-/// (Refs GitHub #2336 AC1 / AC2 / Bottleneck 6).
-const Duration kE2eDefaultCivilianWorkMenuMountTimeout = Duration(seconds: 5);
-
-/// Default phase label emitted by [e2eAwaitCivilianWorkMenuMounted] — matches
-/// the legacy inline `wait_until_civilian_work_menu` phase used by
-/// [e2eTapFirstAssignInCivilianPanel] before the lift. Callers tapping a
-/// title-scoped `Assign` row pass `wait_until_civilian_work_menu_row` to
-/// preserve the historical label split.
-const String kE2eDefaultCivilianWorkMenuMountPhase =
-    'wait_until_civilian_work_menu';
-
-/// Polls until any civilian work-menu label
-/// (one of [kE2eCivilianWorkMenuLabels]) becomes hit-testable, using
-/// [e2eWaitUntilAnyFinderHitTestable] adaptive backoff (25 → 500 ms cap).
-///
-/// Lifted from the inline post-`Assign`-tap waits formerly duplicated in
-/// [e2eTapFirstAssignInCivilianPanel] (`wait_until_civilian_work_menu`) and
-/// [e2eTapAssignOnCivilianRowWithTitle] (`wait_until_civilian_work_menu_row`).
-/// Both call sites now delegate here so the label set, the 5 s default
-/// timeout, and the underlying poll cadence have a single source of truth.
-/// Future callers that tap an `Assign` button (or an upstream affordance that
-/// equivalently mounts the work menu) compose this helper directly without
-/// duplicating the recipe. Refs GitHub #2336 AC1 / AC2 / Bottleneck 6.
-///
-/// Contract:
-///
-/// - Builds the finder list from [kE2eCivilianWorkMenuLabels] in declaration
-///   order so the existential short-circuit inside
-///   [e2eWaitUntilAnyFinderHitTestable] resolves in a deterministic order
-///   (`Build improvement` first), matching the pre-lift inline blocks.
-/// - On timeout, propagates the [TestFailure] raised by
-///   [e2eWaitUntilAnyFinderHitTestable] verbatim — the lift does not change
-///   fail-fast semantics. The accompanying widget-test pin in
-///   `app/test/e2e_await_civilian_work_menu_mounted_test.dart` exercises both
-///   the immediate-found fast path and the timeout path so a regression here
-///   surfaces in PR checks rather than as a silent late-timeout in the Linux
-///   integration suite.
-/// - Emits perf timings via the inner [e2eWaitUntilAnyFinderHitTestable] call
-///   only — no extra `wait_until_civilian_work_menu_*` counter so repeated
-///   waits in a single scenario are still attributable to the existing
-///   `wait_until_any_calls` counter.
-Future<void> e2eAwaitCivilianWorkMenuMounted(
-  WidgetTester tester, {
-  Duration timeout = kE2eDefaultCivilianWorkMenuMountTimeout,
-  String phaseName = kE2eDefaultCivilianWorkMenuMountPhase,
-  E2ePerfLog? perf,
-}) async {
-  final finders = <Finder>[
-    for (final label in kE2eCivilianWorkMenuLabels) find.text(label),
-  ];
-  await e2eWaitUntilAnyFinderHitTestable(
-    tester,
-    finders,
-    timeout: timeout,
-    perf: perf,
-    phaseName: phaseName,
-  );
-}
-
-/// Taps the first visible **Assign** in the civilian panel work menu (GitHub #2336 H9).
-///
-/// Settles + taps the resolved Assign via [e2eEnsureVisibleAndTapHitTestable]
-/// so the call site picks up the shared `ensureVisible` + `hitTestable()`
-/// resolve recipe (the same path the panel-opener rail/marker taps already
-/// use). The inline `await tester.pump();` between `ensureVisible` and
-/// `tap` is removed because the lifted helper does not insert a settle
-/// frame; the shared helper instead resolves `hitTestable()` first and
-/// falls back to the raw finder, which is the canonical AC5 / AC10
-/// adaptive-pre-tap recipe documented in the e2e-ui-stability rule
-/// (Refs GitHub #2336 AC5 / AC10 / Bottleneck 6).
-/// Taps the enclosing [CtNinePatchButton] for a button [label] finder, falling
-/// back to the shared [e2eEnsureVisibleAndTapHitTestable] label recipe when no
-/// `CtNinePatchButton` ancestor resolves (for example a plain `InkWell` work-
-/// menu row from `_showOrderMenu`).
-///
-/// Why this exists (Refs GitHub #2336 AC6 / AC10): `CtNinePatchButton`
-/// (`#2859` R1) renders its label inside a `Stack` whose top child is a
-/// `Positioned.fill(IgnorePointer(_BrassCornerBrackets))` painted over the
-/// engraved-label content, wrapped by a `Material` + `InkWell`. On headless
-/// Linux a `tester.tap` aimed at the inner `Text` center resolves to the label
-/// glyph rather than the `InkWell`'s gesture region, so the `onPressed`
-/// callback (e.g. the civilian-row **Assign** → `_showOrderMenu`) never fires
-/// and the downstream work-menu wait times out. Resolving the tap to the
-/// enclosing button — the canonical interactive control per
-/// `colonizethis-e2e-ui-stability.mdc` (*scope interactions to the actual
-/// control*) — fires `onPressed` deterministically.
-///
-/// Returns `true` when a tap was issued (button ancestor or label fallback),
-/// `false` only when [label] resolves to zero elements.
-Future<bool> e2eTapEnclosingNinePatchButtonOrLabel(
-  WidgetTester tester,
-  Finder label,
-) async {
-  if (label.evaluate().isEmpty) {
-    return false;
-  }
-  final button = find.ancestor(
-    of: label,
-    matching: find.byType(CtNinePatchButton),
-  );
-  if (button.evaluate().isEmpty) {
-    return e2eEnsureVisibleAndTapHitTestable(tester, label);
-  }
-  final target = button.first;
-  await e2eScrollButtonIntoHitTestableView(tester, target);
-  final hit = target.hitTestable();
-  final resolved = hit.evaluate().isNotEmpty ? hit.first : target;
-  await tester.tap(resolved, warnIfMissed: false);
-  return true;
-}
-
-/// Best-effort scrolls [button] until it is genuinely **hit-testable**
-/// (visible inside the render-view bounds), not merely *built*.
-///
-/// Why this exists (Refs GitHub #2336 AC6 / AC7 / AC10): a `ListView` keeps
-/// rows just outside the viewport laid out (within `cacheExtent`), so they are
-/// neither `Offstage` nor removed from the element tree. As a result both
-/// `WidgetController.scrollUntilVisible` (which stops as soon as the finder
-/// `evaluate()` is non-empty) and a single `tester.ensureVisible` can no-op on
-/// a row whose center is still below the render view — e.g. the first civilian
-/// **Assign** button laid out at `y≈819` inside a `1280×720` headless view. A
-/// `tester.tap` then derives an off-screen offset, misses the hit test, never
-/// fires `onPressed`, and the downstream work-menu wait times out. Driving the
-/// enclosing `Scrollable` until the button is hit-testable makes the tap land
-/// deterministically per `colonizethis-e2e-ui-stability.mdc`
-/// (*verify visibility before interaction*).
-Future<void> e2eScrollButtonIntoHitTestableView(
-  WidgetTester tester,
-  Finder button,
-) async {
-  if (button.hitTestable().evaluate().isNotEmpty) {
-    return;
-  }
-  try {
-    await tester.ensureVisible(button);
-    await tester.pump();
-  } catch (_) {}
-  if (button.hitTestable().evaluate().isNotEmpty) {
-    return;
-  }
-  final scrollable = find.ancestor(
-    of: button,
-    matching: find.byType(Scrollable),
-  );
-  if (scrollable.evaluate().isEmpty) {
-    return;
-  }
-  const maxScrollSteps = 20;
-  for (
-    var i = 0;
-    i < maxScrollSteps && button.hitTestable().evaluate().isEmpty;
-    i++
-  ) {
-    final dragged = await e2eDragScrollableFromVisiblePoint(
-      tester,
-      scrollable,
-      const Offset(0, -120),
-    );
-    if (!dragged) {
-      break;
-    }
-    await tester.pump(const Duration(milliseconds: 50));
-  }
-}
-
-/// Logical size of the test render view (for example `1280×720` on the headless
-/// Linux desktop host). Used to clamp synthetic gesture start points inside the
-/// visible bounds.
-Size e2eRenderViewSize(WidgetTester tester) {
-  return tester.view.physicalSize / tester.view.devicePixelRatio;
-}
-
-/// Drags [scrollable] by [delta] starting from a point guaranteed to lie inside
-/// both the scrollable's on-screen extent and the render view, returning `true`
-/// when a drag was issued.
-///
-/// Why this exists (Refs GitHub #2336 AC6 / AC7 / AC10): a freshly opened
-/// bottom-sheet panel can still be mid entrance-animation, so the `Scrollable`'s
-/// geometric center may sit below the render view (observed at `y≈918` in a
-/// `1280×720` headless view). A plain `tester.drag(scrollable, ...)` derives its
-/// start offset from that off-screen center, misses the hit test, and scrolls
-/// nothing — so rows below the fold stay unreachable. Grabbing the on-screen
-/// portion of the scrollable instead drives the list deterministically per
-/// `colonizethis-e2e-ui-stability.mdc` (*verify visibility before interaction*).
-/// Returns `false` when no usable on-screen portion of the scrollable exists.
-Future<bool> e2eDragScrollableFromVisiblePoint(
-  WidgetTester tester,
-  Finder scrollable,
-  Offset delta,
-) async {
-  if (scrollable.evaluate().isEmpty) {
-    return false;
-  }
-  final rect = tester.getRect(scrollable.first);
-  final view = e2eRenderViewSize(tester);
-  final top = rect.top < 0 ? 0.0 : rect.top;
-  final bottom = rect.bottom > view.height ? view.height : rect.bottom;
-  if (bottom - top < 8.0) {
-    return false;
-  }
-  final startX = rect.center.dx.clamp(8.0, view.width - 8.0).toDouble();
-  final startY = ((top + bottom) / 2).clamp(8.0, view.height - 8.0).toDouble();
-  await tester.dragFrom(Offset(startX, startY), delta);
-  return true;
-}
-
-/// Pumps until [scrollable] has a meaningful on-screen extent (a freshly opened
-/// panel has finished sliding up) or [timeout] elapses.
-///
-/// Guards interactions that target a bottom-sheet panel immediately after it is
-/// opened: without this settle the panel's `Scrollable` can still be animating
-/// up from the bottom edge, leaving its content below the render view (see
-/// [e2eDragScrollableFromVisiblePoint]). Refs GitHub #2336 AC6 / AC7 / AC10.
-Future<void> e2eWaitScrollableOnScreen(
-  WidgetTester tester,
-  Finder scrollable, {
-  Duration timeout = const Duration(seconds: 5),
-  String phaseName = 'pump_until_panel_scrollable_onscreen',
-}) async {
-  await e2ePumpUntilConditionOrIdle(
-    tester,
-    () {
-      if (scrollable.evaluate().isEmpty) {
-        return false;
-      }
-      final rect = tester.getRect(scrollable.first);
-      final view = e2eRenderViewSize(tester);
-      final top = rect.top < 0 ? 0.0 : rect.top;
-      final bottom = rect.bottom > view.height ? view.height : rect.bottom;
-      return bottom - top > 40.0;
-    },
-    timeout: timeout,
-    phaseName: phaseName,
-  );
-}
-
-double _e2eScrollPositionPixels(WidgetTester tester, Finder scrollable) {
-  final state = tester.state<ScrollableState>(scrollable.first);
-  return state.position.pixels;
-}
-
-/// Scrolls the civilian-panel [ListView] through its full extent (bottom then
-/// top) so every unit row is built before [e2eCollectTextPreorder] walks the
-/// tree.
-///
-/// Why this exists (Refs GitHub #2336 AC6 / AC7 / AC10): the bottom-sheet
-/// panel caps list height (~308 dp visible) while the E2E roster can list more
-/// units than fit on screen. `ListView` virtualization omits off-screen row
-/// `Text` nodes, so a preorder snapshot taken at scroll offset `0` can be
-/// shorter than [civilianUnitsPanelExpectedTexts] (for example the second
-/// idle `Explorer` row at indices 25–29). Driving the list to max extent and
-/// back — together with the E2E `cacheExtent` on [UnitsPanelShell] — keeps
-/// every row mounted for the assertion pass.
-Future<void> e2ePrepareCivilianPanelListForTextCollection(
-  WidgetTester tester,
-) async {
-  final root = find.byKey(kCtE2ECivilianPanelRootKey);
-  if (root.evaluate().isEmpty) {
-    return;
-  }
-  final scrollable = find.descendant(
-    of: root,
-    matching: find.byType(Scrollable),
-  );
-  if (scrollable.evaluate().isEmpty) {
-    return;
-  }
-  await e2eWaitScrollableOnScreen(tester, scrollable);
-  const maxSteps = 25;
-  for (var i = 0; i < maxSteps; i++) {
-    final before = _e2eScrollPositionPixels(tester, scrollable);
-    final dragged = await e2eDragScrollableFromVisiblePoint(
-      tester,
-      scrollable,
-      const Offset(0, -200),
-    );
-    await tester.pump(const Duration(milliseconds: 50));
-    final after = _e2eScrollPositionPixels(tester, scrollable);
-    if (!dragged || (after - before).abs() < 0.5) {
-      break;
-    }
-  }
-  for (var i = 0; i < maxSteps; i++) {
-    final before = _e2eScrollPositionPixels(tester, scrollable);
-    if (before < 0.5) {
-      break;
-    }
-    await e2eDragScrollableFromVisiblePoint(
-      tester,
-      scrollable,
-      const Offset(0, 200),
-    );
-    await tester.pump(const Duration(milliseconds: 50));
-    final after = _e2eScrollPositionPixels(tester, scrollable);
-    if ((after - before).abs() < 0.5) {
-      break;
-    }
-  }
-}
-
-Future<void> e2eTapFirstAssignInCivilianPanel(WidgetTester tester) async {
-  final root = find.byKey(kCtE2ECivilianPanelRootKey);
-  final listView = find.descendant(of: root, matching: find.byType(ListView));
-  expect(listView, findsOneWidget);
-  final panelScrollable = find.descendant(
-    of: listView,
-    matching: find.byType(Scrollable),
-  );
-  expect(panelScrollable, findsOneWidget);
-  final assign = find.descendant(of: root, matching: find.text('Assign'));
-  expect(assign, findsWidgets);
-  final firstAssign = assign.first;
-  await tester.scrollUntilVisible(
-    firstAssign,
-    120,
-    scrollable: panelScrollable,
-  );
-  final didTap = await e2eTapEnclosingNinePatchButtonOrLabel(
-    tester,
-    firstAssign,
-  );
-  expect(
-    didTap,
-    isTrue,
-    reason:
-        'First Assign in civilian panel resolved to zero elements after the '
-        'findsWidgets guard passed; the downstream work-menu wait would race '
-        'a not-yet-tapped Assign button.',
-  );
-  // Lifted post-tap work-menu wait: see [e2eAwaitCivilianWorkMenuMounted] for
-  // the canonical label set and 5 s timeout. The default phase name preserves
-  // the legacy inline `wait_until_civilian_work_menu` label byte-for-byte.
-  // Refs GitHub #2336 AC1 / AC2 / Bottleneck 6.
-  await e2eAwaitCivilianWorkMenuMounted(tester);
-}
-
-/// Taps a civilian work-order label (for example `Build improvement`,
-/// `Prospect`, or `Explore`) inside the open civilian-panel work menu, using
-/// [e2eEnsureVisibleAndTapHitTestable] so the tap fires from a canonical
-/// hit-testable position even when the label is rendered slightly off-screen
-/// by a transient overlay or a small surface.
-///
-/// Replaces the legacy raw `tester.tap(find.text(workOrderLabel))` taps in
-/// `new_game_full_turn_e2e_test.dart` that ran right after
-/// [e2eTapFirstAssignInCivilianPanel] / [e2eTapAssignOnCivilianRowWithTitle].
-/// Those tap-Assign helpers only guarantee that *one* of
-/// `{Build improvement, Prospect, Explore}` is hit-testable on return — the
-/// **specific** label the caller wants may still be obscured (e.g. behind a
-/// soft keyboard, a transient bottom sheet, or a viewport that needs scrolling)
-/// at the instant of the tap. Wrapping the tap in this helper enforces the
-/// e2e-ui-stability rule's *verify visibility before interaction* directive
-/// at the same call sites that previously skipped it. Refs GitHub #2336 AC1 /
-/// AC2 / AC10; `colonizethis-e2e-ui-stability.mdc`.
-///
-/// Contract:
-///
-/// - Throws [TestFailure] via [expect] when no `Text` widget matching
-///   [workOrderLabel] is mounted at all, so the surrounding scenario fails
-///   fast at the offending step rather than burning the downstream
-///   work-tile timeout against an unmounted work menu (the same fail-fast
-///   contract documented on [e2eTapFirstAssignInCivilianPanel]).
-/// - Delegates to [e2eEnsureVisibleAndTapHitTestable], which best-effort
-///   scrolls the label into view, prefers the `hitTestable()` resolution,
-///   and falls back to the raw finder when no hit-testable element resolves
-///   (the same path the panel-opener rail/marker taps already use).
-/// - Throws [TestFailure] when [e2eEnsureVisibleAndTapHitTestable] reports
-///   no tap was issued (zero-element resolution after the presence assertion
-///   passes — only possible under unusual race conditions where the label
-///   vanishes between the assertion and the resolve). This keeps the helper
-///   sound for callers that rely on a tap actually firing before the
-///   downstream work-tile pick begins.
-Future<void> e2eTapCivilianWorkOrderLabel(
-  WidgetTester tester,
-  String workOrderLabel,
-) async {
-  final label = find.text(workOrderLabel);
-  expect(
-    label,
-    findsWidgets,
-    reason:
-        'Civilian work-order label "$workOrderLabel" not found in the open '
-        'work menu. Either the preceding Assign tap did not mount the work '
-        'menu, or the label name has drifted from the production scaffold.',
-  );
-  final didTap = await e2eEnsureVisibleAndTapHitTestable(tester, label);
-  expect(
-    didTap,
-    isTrue,
-    reason:
-        'Civilian work-order label "$workOrderLabel" was present but the '
-        'shared ensureVisible/hit-testable tap path issued no tap; the '
-        'downstream work-tile pick would race a not-yet-tapped label.',
-  );
-}
-
-/// Taps **Assign** on the [CivilianUnitRowCard] whose title is exactly
-/// [unitTypeTitle] (GitHub #2336 H9).
-///
-/// The civilian panel migrated its unit rows off Material `ListTile` to the
-/// bespoke [CivilianUnitRowCard] (Refs #2914 S8); this helper therefore scopes
-/// the per-row **Assign** lookup to that card so the row's `CtNinePatchButton`
-/// resolves after the migration instead of silently matching nothing.
-Future<void> e2eTapAssignOnCivilianRowWithTitle(
-  WidgetTester tester,
-  String unitTypeTitle,
-) async {
-  final root = find.byKey(kCtE2ECivilianPanelRootKey);
-  final listView = find.descendant(of: root, matching: find.byType(ListView));
-  expect(listView, findsOneWidget);
-  final panelScrollable = find.descendant(
-    of: listView,
-    matching: find.byType(Scrollable),
-  );
-  expect(panelScrollable, findsOneWidget);
-  final titlesInList = find.descendant(
-    of: listView,
-    matching: find.text(unitTypeTitle),
-  );
-  // The panel may still be mid entrance-animation; settle it on-screen before
-  // dragging so gesture start points land inside the render view (#2336 AC6).
-  await e2eWaitScrollableOnScreen(tester, panelScrollable);
-  final sw = Stopwatch()..start();
-  while (titlesInList.evaluate().isEmpty &&
-      sw.elapsed < const Duration(seconds: 20)) {
-    final dragged = await e2eDragScrollableFromVisiblePoint(
-      tester,
-      panelScrollable,
-      const Offset(0, -120),
-    );
-    if (!dragged) {
-      await e2eWaitScrollableOnScreen(tester, panelScrollable);
-    }
-    await e2ePumpUntilConditionOrIdle(
-      tester,
-      () => titlesInList.evaluate().isNotEmpty,
-      timeout: const Duration(milliseconds: 200),
-      phaseName: 'pump_until_civilian_title_visible_after_scroll_drag',
-    );
-  }
-  expect(
-    titlesInList,
-    findsWidgets,
-    reason:
-        'Timed out scrolling civilian panel for a visible "$unitTypeTitle" row',
-  );
-  final n = titlesInList.evaluate().length;
-  for (var i = 0; i < n; i++) {
-    final titleAt = titlesInList.at(i);
-    try {
-      await tester.scrollUntilVisible(
-        titleAt,
-        120,
-        scrollable: panelScrollable,
-      );
-    } catch (_) {}
-    try {
-      await tester.ensureVisible(titleAt);
-    } catch (_) {}
-    final rowCard = find.ancestor(
-      of: titleAt,
-      matching: find.byType(CivilianUnitRowCard),
-    );
-    final assign = find.descendant(of: rowCard, matching: find.text('Assign'));
-    if (assign.evaluate().isEmpty) {
-      continue;
-    }
-    final assignHit = assign.first;
-    final didTap = await e2eTapEnclosingNinePatchButtonOrLabel(
-      tester,
-      assignHit,
-    );
-    expect(
-      didTap,
-      isTrue,
-      reason:
-          'Assign on civilian row "$unitTypeTitle" resolved to zero elements '
-          'after the per-row Assign descendant guard passed; the downstream '
-          'work-menu wait would race a not-yet-tapped Assign button.',
-    );
-    // Lifted post-tap work-menu wait: see [e2eAwaitCivilianWorkMenuMounted].
-    // The legacy `wait_until_civilian_work_menu_row` phase label is preserved
-    // explicitly so the title-scoped sibling helper stays distinguishable in
-    // perf-timing dumps (Refs GitHub #2336 AC1 / AC2 / Bottleneck 6).
-    await e2eAwaitCivilianWorkMenuMounted(
-      tester,
-      phaseName: 'wait_until_civilian_work_menu_row',
-    );
-    return;
-  }
-  fail('No idle Assign row for unit type "$unitTypeTitle" in civilian panel');
-}
-
-/// Default phase label emitted by [e2eDismissTransientUi] when a caller does
-/// not override [E2ePerfLog] attribution.
-///
-/// The constant is exposed so widget-test pins and downstream perf-marker
-/// scrapers (for example the GitHub #2336 AC8 baseline timing pipeline via
-/// `tool/run_e2e_timing.sh` / `tool/compare_e2e_timing.sh`) can refer to the
-/// canonical label by name instead of hard-coding the literal string. Mirrors
-/// the [kE2eDefaultAdvanceGameStartIntroPhase] / [kE2eDefaultWaitForMapHudPhase]
-/// convention used by sibling helper-level perf attribution.
-const String kE2eDefaultDismissTransientUiPhase = 'dismiss_transient_ui';
-
-/// Dismisses snackbars, generic OK dialogs, [AlertDialog] actions, bottom sheets,
-/// and [CtDialogShell] overlays (union of fleet + full-turn E2E paths).
-///
-/// Perf attribution (Refs GitHub #2336 AC8 / baseline measurement):
-///
-/// - When [perf] is non-null, emits exactly one `E2E_TIMING|phase=[phaseName]`
-///   line per call with a `result=...` meta tag that names the branch that
-///   handled (or short-circuited through) the dispatch waterfall:
-///   `result=intro_advanced` (intro overlay blocker dispatched to
-///   [e2eAdvanceGameStartIntroUntilDismissed]), `result=snackbar`
-///   ([e2eDismissSnackBarIfPresent] returned true),
-///   `result=generic_ok` ([e2eDismissGenericOkIfPresent] returned true),
-///   `result=alert_dialog` ([e2eDismissAlertDialogIfPresent] returned true),
-///   and `result=broad_sweep` (no early-return branch fired, so the function
-///   ran the [BottomSheet] close attempt and the
-///   [e2eDismissCtDialogShellBroadSweepIfPresent] tail). The default phase
-///   label is [kE2eDefaultDismissTransientUiPhase]; callers may pass a
-///   different [phaseName] to keep distinct dispatch sites separable in
-///   perf-timing dumps.
-/// - With `perf: null` (the default the legacy widget-test pins and any
-///   opt-out callers still pass) the helper emits no
-///   `E2E_TIMING|phase=[phaseName]` line for the dispatcher itself. Inner
-///   helpers retain their own attribution contracts when callers thread a
-///   non-null `perf` through.
-///
-/// The `dismiss_transient_ui_calls` counter still bumps once on entry,
-/// regardless of [perf] phase, so legacy log scrapers that count dispatch
-/// invocations remain stable.
-Future<void> e2eDismissTransientUi(
-  WidgetTester tester, {
-  E2ePerfLog? perf,
-  String phaseName = kE2eDefaultDismissTransientUiPhase,
-}) async {
-  perf?.bumpCounter('dismiss_transient_ui_calls');
-  final sw = Stopwatch()..start();
-  if (e2eGameStartIntroBlocksUi(tester)) {
-    await e2eAdvanceGameStartIntroUntilDismissed(tester, perf: perf);
-    perf?.timing(phaseName, sw.elapsed, meta: 'result=intro_advanced');
-    return;
-  }
-  // Shared SnackBar dismissal: lifted into [e2eDismissSnackBarIfPresent]
-  // so the hit-testable-action tap recipe is single-source-of-truth and
-  // pinned by widget tests. The pre-lift inline block checked
-  // `snackAction.hitTestable()` for presence but tapped `snackAction.first`
-  // (the first [TextButton] without the hit-testable filter), which could
-  // miss the tap when the first [TextButton] was covered by another overlay
-  // while a later one remained hit-testable. The lifted form taps the
-  // hit-testable filter's first match — matching the adjacent AlertDialog
-  // and CtDialogShell branches below that already use the filtered finder
-  // for both check and tap. Refs GitHub #2336 AC1 / AC2 / AC10.
-  if (await e2eDismissSnackBarIfPresent(tester, perf: perf)) {
-    perf?.timing(phaseName, sw.elapsed, meta: 'result=snackbar');
-    return;
-  }
-  // Shared top-level OK dismissal: lifted into [e2eDismissGenericOkIfPresent]
-  // so the legacy `find.text('OK').hitTestable()` tap + 2 s
-  // `e2ePumpUntilFinderEmpty` recipe is single-source-of-truth and pinned by
-  // widget tests. The pre-lift inline block lived between the SnackBar and
-  // AlertDialog branches and targeted a top-level OK button (not nested in
-  // an [AlertDialog]) — a stray confirmation banner above the map HUD
-  // between phases. The lifted form preserves the legacy English literal
-  // (`OK`), the 2 s dismiss budget, and the post-tap return semantics.
-  // Refs GitHub #2336 AC1 / AC2 / Bottleneck 6.
-  if (await e2eDismissGenericOkIfPresent(tester, perf: perf)) {
-    perf?.timing(phaseName, sw.elapsed, meta: 'result=generic_ok');
-    return;
-  }
-  // Shared AlertDialog dismissal: lifted into [e2eDismissAlertDialogIfPresent]
-  // so the labelled-button-priority tap + `handlePopRoute` fallback recipe is
-  // single-source-of-truth and pinned by widget tests. The lifted form
-  // preserves the legacy English label priority (`Close` → `OK` → `Cancel` →
-  // `Yes`), the 2 s dismiss budget, and the unconditional `handlePopRoute`
-  // fallback when no labelled button is hit-testable. Refs GitHub #2336
-  // AC1 / AC2 / Bottleneck 6.
-  if (await e2eDismissAlertDialogIfPresent(tester, perf: perf)) {
-    perf?.timing(phaseName, sw.elapsed, meta: 'result=alert_dialog');
-    return;
-  }
-  if (find.byType(BottomSheet).evaluate().isNotEmpty) {
-    await e2eCloseBottomSheet(tester, perf: perf);
-  }
-  // Shared CtDialogShell broad-sweep dismissal: lifted into
-  // [e2eDismissCtDialogShellBroadSweepIfPresent] so the English-only
-  // close-candidate sweep (`Cancel` → `Close` → `Icons.close` →
-  // `Icons.arrow_back`) plus the `tester.binding.handlePopRoute()` fallback
-  // are single-source-of-truth and pinned by widget tests. After this lift,
-  // every overlay branch of the broad-spectrum sweep delegates to a focused
-  // shared helper — no inline dismissal recipes remain in this function's
-  // overlay branches. Refs GitHub #2336 AC1 / AC2 / Bottleneck 6.
-  await e2eDismissCtDialogShellBroadSweepIfPresent(tester, perf: perf);
-  // Fall-through dispatch attribution: even when no early-return branch
-  // fired (snackbar / generic OK / alert dialog), the helper still runs the
-  // BottomSheet close attempt and the CtDialogShell broad-sweep tail before
-  // returning. Reporting `result=broad_sweep` keeps the AC8 timing pipeline
-  // able to separate fast labelled-branch dispatches from the broad-sweep
-  // path even when neither inner attempt finds an overlay to close (those
-  // inner helpers emit their own timing only when they actually run their
-  // wait loops).
-  perf?.timing(phaseName, sw.elapsed, meta: 'result=broad_sweep');
 }
 
 /// Pumps until [finder] matches at least one widget or [timeout] elapses,
@@ -1363,94 +386,4 @@ void Function(String step) e2eMakeWallClockGuard({
       );
     }
   };
-}
-
-/// Returns a [Finder] matching every `RadioListTile<…>` widget that is a
-/// descendant of any currently-mounted [AlertDialog].
-///
-/// Contract (issue #2336 AC1 / AC2):
-///
-///   - Composes `find.descendant(of: find.byType(AlertDialog), matching: …)`
-///     so the search is scoped to dialog subtrees only — `RadioListTile`s
-///     outside an [AlertDialog] (panels, settings sheets, etc.) are never
-///     returned. Naval / civilian / production move dialogs all surface
-///     destination tiles via `RadioListTile<…>`; scoping to [AlertDialog]
-///     keeps the move-segment helpers from accidentally hitting a panel
-///     radio.
-///   - The matcher inspects [Widget.runtimeType] via `toString()` so it
-///     accepts any generic instantiation
-///     (`RadioListTile<String>`, `RadioListTile<int>`, etc.) without
-///     pulling in the concrete type argument. The fleet-reach move dialog
-///     parameterises radio tiles on the destination string id; binding to
-///     `RadioListTile<String>` directly would silently miss future
-///     dialogs that switch to a different parameter type.
-///   - Pure — the function reads no globals, returns a new [Finder] on
-///     every call, and never throws. The returned [Finder] is lazy and
-///     only resolves against the active widget tree when iterated, which
-///     keeps it safe to construct outside an `await tester.pump()`
-///     boundary.
-///
-/// Mirrors the lifted public-name pattern used by
-/// `e2eNonHomeHumanFleetInNewWorldFromCtSnapshot` and
-/// `e2eNavalPanelShowsNonHomeFleetInNewWorld` so callers consume the
-/// AC1 barrel (`e2e_helpers.dart`) only. A silent rename or accidental
-/// scope-removal (matching every `RadioListTile` across the tree)
-/// would re-introduce false positives in move-segment dialogs that
-/// also host non-destination radios on the surrounding screen
-/// (Refs `SPEC/program/e2e-integration-tests.md` § Determinism).
-Finder e2eRadioListTilesInAlertDialogs() {
-  return find.descendant(
-    of: find.byType(AlertDialog),
-    matching: find.byWidgetPredicate(
-      (w) => w.runtimeType.toString().startsWith('RadioListTile<'),
-    ),
-  );
-}
-
-/// Returns a [Finder] matching the move-fleet dialog container, tolerating
-/// **either** a Material [AlertDialog] **or** the production [CtDialogShell].
-///
-/// The production `MoveFleetDialog` renders as a [CtDialogShell] (a Material
-/// `Dialog`, **not** an `AlertDialog`) per `SPEC/ui/move-fleet-dialog.md` and
-/// the catalog Material-design ban. The fleet-reach move helpers were written
-/// against the legacy `AlertDialog` shape; this finder lets the same helpers
-/// drive the live `CtDialogShell` dialog while the focused widget-test pins —
-/// which still mount `AlertDialog` fixtures — keep validating the helper logic
-/// (Refs #2336; first-fleet-move parity with `e2eAttemptFirstFleetMoveOrCancel`).
-Finder e2eMoveFleetDialogFinder() {
-  return find.byWidgetPredicate((w) => w is AlertDialog || w is CtDialogShell);
-}
-
-/// Returns a [Finder] matching every selectable destination row inside the
-/// active move-fleet dialog ([e2eMoveFleetDialogFinder]).
-///
-/// Tolerates both dialog shapes (Refs #2336):
-///
-///   - **Production** ([CtDialogShell]): custom `_MoveFleetDestinationRow`
-///     widgets, each keyed via [kCtE2EMoveFleetDestinationRowKey] under
-///     `CT_E2E`. No Material `Radio` / `RadioListTile` is used, per
-///     `SPEC/ui/move-fleet-dialog.md` § Layout.
-///   - **Widget-test fixtures** ([AlertDialog]): legacy `RadioListTile<…>`
-///     rows (matched on `runtimeType.toString()` so any generic
-///     instantiation qualifies).
-///
-/// The finder is dialog-scoped so panel-side rows never leak in, and returns a
-/// fresh lazy [Finder] on every call. The fleet-reach picker
-/// ([e2ePickMoveDestinationAndConfirm]) taps `.first` when no warp row is
-/// present.
-Finder e2eMoveFleetDestinationRows() {
-  return find.descendant(
-    of: e2eMoveFleetDialogFinder(),
-    matching: find.byWidgetPredicate((w) {
-      final Key? key = w.key;
-      if (key is ValueKey &&
-          key.value is String &&
-          (key.value as String).startsWith(
-            kCtE2EMoveFleetDestinationRowKeyPrefix,
-          )) {
-        return true;
-      }
-      return w.runtimeType.toString().startsWith('RadioListTile<');
-    }),
-  );
 }
